[Chore]#111-폴더변경 및 파일명 변경
